added check that the item exists before trying to delete it, otherwise returns false.
git-svn-id: https://taverna.googlecode.com/svn/taverna/engine/net.sf.taverna.t2.core/trunk@8879 bf327186-88b3-11dd-a302-d386e5130c1c

+++ reference-impl/src/main/java/net/sf/taverna/t2/reference/impl/ErrorDocumentServiceImpl.java
@@ -153,6 +153,7 @@ public class ErrorDocumentServiceImpl extends AbstractErrorDocumentServiceImpl
 			throws ReferenceServiceException {
 		checkDao();
 		ErrorDocument doc = errorDao.get(reference);
+		if (doc==null) return false;
 		return errorDao.delete(doc);
 	}
